+++ R/raw.getNoSamples.R
@@ -1,13 +1,13 @@
 #' returns approximate number of samples that were made
 #'
 #' @param pfad path to the RAW folder
-#' @param underscoreComments include files with comments that have _
+#' @param recursive if \code{TRUE} search also subfolders
 #' @return number of samples that were made with the NTE
 #' @examples
 #' raw.getNoSamples(raw.getSamplePath())
 #'
 #' @export
-raw.getNoSamples <- function(pfad, underscoreComments=TRUE) {
-  file.list = raw.findFiles(pfad, instrument='NTE', underscoreComments)
-  length(file.list)
+raw.getNoSamples <- function(pfad, recursive=FALSE) {
+  q = raw.inspectFolder(pfad, recursive=recursive)
+  length(strsplit(q$samples,",")[[1]])
 }