+++ R/raw.splitFilename.R
@@ -8,21 +8,11 @@ strsplit.max <- function(x,split,maxNo) {
   s
 }
 
-# empty.table = data.frame(
-#   Date = '',
-#   Date.formatted = as.Date('19900101', format='%Y%m%d'),
-#   Project = '',
-#   User = '',
-#   Instrument = NA,
-#   Sample = '',
-#   Description = '',
-#   Extension = '',
-#   Filename = ''
-# )
 
 #' splits filename into fields
 #'
-#' @param filename filename in RAW folder format
+#' @param filelist filename in RAW folder format
+#' @importFrom tools file_ext
 #' @return data frame
 #' @examples
 #' raw.splitFilename('20010101-project-user-inst-sample.DAT')
@@ -35,45 +25,4 @@ raw.splitFilename <- function(filelist) {
   d1$Extension = tools::file_ext(filelist)
   d1
 }
-  # d1 = as.data.frame(t(sapply(filelist, FUN=function(x) { strsplit.max(x,'_',6) } )))
-  # s = strsplit.max(filename,'_',6)
-  # if (length(s)<6) {
-  #   e = empty.table
-  #   e$Filename = filename
-  #   if (length(s)==5) {
-  #     Extension = gsub('.*\\.([^.]+)$','\\1',s[5])
-  #     sample =  gsub('(.*)\\.[^.]+$','\\1',s[5])
-  #     q = strsplit.max(sample,'-',2)
-  #     if(length(q)==2) {
-  #       s[5] = q[1]
-  #       desc = paste0(q[2],'.',Extension)
-  #     } else {
-  #       s[5] = sample
-  #       desc = paste0('.',Extension)
-  #     }
-  #     e = data.frame(
-  #       Date = s[1],
-  #       Date.formatted = as.Date(s[1], format='%Y%m%d'),
-  #       Project = s[2],
-  #       User = s[3],
-  #       Instrument = s[4],
-  #       Sample = s[5],
-  #       Description = gsub('(.*)\\.[^.]+$','\\1',desc),
-  #       Extension = gsub('.*\\.([^.]+)$','\\1',desc),
-  #       Filename = filename
-  #     )
-  #   }
-  #   return(e)
-  # }
-  # data.frame(
-  #   Date = s[1],
-  #   Date.formatted = as.Date(s[1], format='%Y%m%d'),
-  #   Project = s[2],
-  #   User = s[3],
-  #   Instrument = s[4],
-  #   Sample = s[5],
-  #   Description = gsub('(.*)\\.[^.]+$','\\1',s[6]),
-  #   Extension = gsub('.*\\.([^.]+)$','\\1',s[6]),
-  #   Filename = filename
-  # )
-# }
+