+++ _main.R
@@ -3,3 +3,6 @@
 
 devtools::document()
 pkgdown::build_site()
+
+library(covr)
+report()