+++ tests/testthat/test-checkTestFiles.R
@@ -15,15 +15,19 @@ test_that("Find invalid files.", {
 })
 
 
+test_that("Verify MD5 strings for files.", {
+  q = raw.getNoSamples(raw.getSamplePath())
+  expect_equal(q, 3)
+})
+
 
 test_that("Verify MD5 strings for files.", {
   md5str = raw.getMD5str(raw.getSampleFiles())
   expect_equal(md5str, "7545cd,7545cd,24fefa,7545cd")
 })
 
-# test_that("Add instrument name", {
-#   raw.fixInvalidFile(f, instrument='vsm')
-# })
+
+
 
 test_that("Add instrument name", {
   f="20211213_myProj_MM_MM211202B_IrMn_measure.png"